Optionally remove http version from request lines
Requests in .http files designed for VSCode sometimes (always?)
end the request lines with the http version, e.g. "HTTP/1.1".

However, url.el unconditionally appends HTTP/1.1 to the end of the
urls it's asked to retrieve (and does not support other
versions). This may confuse servers, who end up seeing the version
twice, frequently interpreting the first as part of the query string.
So add an option to strip the http version (if any) from the end of
the request lines in restclient.  Default off to avoid breaking
backwards compatibility.